requirements.txt
djangorestframework_simplejwt==5.5.1
djangorestframework==3.16.1
django-cors-headers==4.7.0
Django==5.2.5
python-decouple==3.8